pyproject.toml
[project]
name = "expensify-os"
version = "0.1.0"
description = "Automated expense management — fetch invoices, download receipts, submit to Expensify"
readme = "README.md"
license = "MIT"
authors = [
    { name = "bultot", email = "robin@bultot.nl" }
]
requires-python = ">=3.13"

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "playwright>=1.49",
    "tenacity>=9.0",
    "structlog>=24.4",
]

[project.scripts]
expensify-os = "expensify_os.cli:cli"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "respx>=0.22",
    "pytest-mock>=3.14",
]
